+++ src/main/java/fr/alasdiablo/janoeo/world/gen/feature/OresFeatures.java
@@ -7,8 +7,6 @@ import fr.alasdiablo.janoeo.init.*;
 import fr.alasdiablo.janoeo.config.FrequencyConfig;
 import net.minecraft.block.Blocks;
 import net.minecraft.world.gen.feature.OreFeatureConfig;
-import net.minecraft.world.gen.feature.template.BlockMatchRuleTest;
-import net.minecraft.world.gen.feature.template.RuleTest;
 
 import java.util.Objects;
 
@@ -28,7 +26,9 @@ public class OresFeatures {
     public static void initNetherDenseOre() {
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_COAL_NETHER_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    Objects.requireNonNull(RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_COAL_NETHER_ORE.getRegistryName(), String.valueOf(i))),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_COAL_NETHER_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     NetherOresBlocks.COAL_NETHER_ORE.getDefaultState(),
                     NetherDenseOresBlocks.DENSE_COAL_NETHER_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_COAL_NETHER_ORE_COUNT.get(),
@@ -38,7 +38,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_DIAMOND_NETHER_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_DIAMOND_NETHER_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_DIAMOND_NETHER_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     NetherOresBlocks.DIAMOND_NETHER_ORE.getDefaultState(),
                     NetherDenseOresBlocks.DENSE_DIAMOND_NETHER_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_DIAMOND_NETHER_ORE_COUNT.get(),
@@ -48,7 +50,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_EMERALD_NETHER_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_EMERALD_NETHER_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_EMERALD_NETHER_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     NetherOresBlocks.EMERALD_NETHER_ORE.getDefaultState(),
                     NetherDenseOresBlocks.DENSE_EMERALD_NETHER_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_EMERALD_NETHER_ORE_COUNT.get(),
@@ -58,7 +62,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_GOLD_NETHER_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_GOLD_NETHER_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_GOLD_NETHER_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     NetherOresBlocks.GOLD_NETHER_ORE.getDefaultState(),
                     NetherDenseOresBlocks.DENSE_GOLD_NETHER_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_GOLD_NETHER_ORE_COUNT.get(),
@@ -68,7 +74,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_IRON_NETHER_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_IRON_NETHER_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_IRON_NETHER_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     NetherOresBlocks.IRON_NETHER_ORE.getDefaultState(),
                     NetherDenseOresBlocks.DENSE_IRON_NETHER_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_IRON_NETHER_ORE_COUNT.get(),
@@ -78,7 +86,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_LAPIS_NETHER_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_LAPIS_NETHER_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_LAPIS_NETHER_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     NetherOresBlocks.LAPIS_NETHER_ORE.getDefaultState(),
                     NetherDenseOresBlocks.DENSE_LAPIS_NETHER_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_LAPIS_NETHER_ORE_COUNT.get(),
@@ -88,7 +98,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_REDSTONE_NETHER_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_REDSTONE_NETHER_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_REDSTONE_NETHER_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     NetherOresBlocks.REDSTONE_NETHER_ORE.getDefaultState(),
                     NetherDenseOresBlocks.DENSE_REDSTONE_NETHER_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_REDSTONE_NETHER_ORE_COUNT.get(),
@@ -98,7 +110,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_QUARTZ_NETHER_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_QUARTZ_NETHER_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(NetherDenseOresBlocks.DENSE_QUARTZ_NETHER_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     Blocks.NETHER_QUARTZ_ORE.getDefaultState(),
                     NetherDenseOresBlocks.DENSE_QUARTZ_NETHER_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_QUARTZ_NETHER_ORE_COUNT.get(),
@@ -113,7 +127,9 @@ public class OresFeatures {
     public static void initDenseOre() {
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_COAL_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_COAL_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_COAL_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     Blocks.COAL_ORE.getDefaultState(),
                     OverworldDenseOresBlocks.DENSE_COAL_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_COAL_ORE_COUNT.get(),
@@ -123,7 +139,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_DIAMOND_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_DIAMOND_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_DIAMOND_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     Blocks.DIAMOND_ORE.getDefaultState(),
                     OverworldDenseOresBlocks.DENSE_DIAMOND_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_DIAMOND_ORE_COUNT.get(),
@@ -133,7 +151,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_EMERALD_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_EMERALD_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_EMERALD_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     Blocks.EMERALD_ORE.getDefaultState(),
                     OverworldDenseOresBlocks.DENSE_EMERALD_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_EMERALD_ORE_COUNT.get(),
@@ -143,7 +163,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_GOLD_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_GOLD_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_GOLD_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     Blocks.GOLD_ORE.getDefaultState(),
                     OverworldDenseOresBlocks.DENSE_GOLD_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_GOLD_ORE_COUNT.get(),
@@ -153,7 +175,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_IRON_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_IRON_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_IRON_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     Blocks.IRON_ORE.getDefaultState(),
                     OverworldDenseOresBlocks.DENSE_IRON_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_IRON_ORE_COUNT.get(),
@@ -163,7 +187,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_LAPIS_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_LAPIS_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_LAPIS_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     Blocks.LAPIS_ORE.getDefaultState(),
                     OverworldDenseOresBlocks.DENSE_LAPIS_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_LAPIS_ORE_COUNT.get(),
@@ -173,7 +199,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_REDSTONE_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_REDSTONE_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_REDSTONE_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     Blocks.REDSTONE_ORE.getDefaultState(),
                     OverworldDenseOresBlocks.DENSE_REDSTONE_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_REDSTONE_ORE_COUNT.get(),
@@ -183,7 +211,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_TIN_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_TIN_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_TIN_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     OverworldOresBlocks.TIN_ORE.getDefaultState(),
                     OverworldDenseOresBlocks.DENSE_TIN_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_TIN_ORE_COUNT.get(),
@@ -193,7 +223,9 @@ public class OresFeatures {
 
         for (int i = 0; i < FREQUENCY_CONFIG.DENSE_COPPER_ORE_MULTIPLIER_FACTOR.get(); i++)
             WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                    RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_COPPER_ORE.getRegistryName(), String.valueOf(i)),
+                    Objects.requireNonNull(
+                            RegistryHelper.setPrefixOnRegistryName(OverworldDenseOresBlocks.DENSE_COPPER_ORE.getRegistryName(), String.valueOf(i))
+                    ),
                     OverworldOresBlocks.COPPER_ORE.getDefaultState(),
                     OverworldDenseOresBlocks.DENSE_COPPER_ORE.getDefaultState(),
                     FREQUENCY_CONFIG.DENSE_COPPER_ORE_COUNT.get(),
@@ -207,7 +239,9 @@ public class OresFeatures {
      */
     public static void initTheEnd() {
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                EndOresBlocks.COAL_END_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        EndOresBlocks.COAL_END_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.END_STONE.get(),
                 EndOresBlocks.COAL_END_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.COAL_THEEND_ORE_SIZE.get(),
@@ -217,7 +251,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                EndOresBlocks.DIAMOND_END_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        EndOresBlocks.DIAMOND_END_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.END_STONE.get(),
                 EndOresBlocks.DIAMOND_END_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.DIAMOND_THEEND_ORE_SIZE.get(),
@@ -227,7 +263,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                EndOresBlocks.EMERALD_END_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        EndOresBlocks.EMERALD_END_ORE.getRegistryName()
+                ),
                 Blocks.END_STONE.getDefaultState(),
                 EndOresBlocks.EMERALD_END_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.EMERALD_THEEND_ORE_COUNT.get(),
@@ -236,7 +274,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                EndOresBlocks.GOLD_END_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        EndOresBlocks.GOLD_END_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.END_STONE.get(),
                 EndOresBlocks.GOLD_END_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.GOLD_THEEND_ORE_SIZE.get(),
@@ -246,7 +286,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                EndOresBlocks.IRON_END_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        EndOresBlocks.IRON_END_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.END_STONE.get(),
                 EndOresBlocks.IRON_END_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.IRON_THEEND_ORE_SIZE.get(),
@@ -256,7 +298,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                EndOresBlocks.LAPIS_END_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        EndOresBlocks.LAPIS_END_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.END_STONE.get(),
                 EndOresBlocks.LAPIS_END_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.LAPIS_THEEND_ORE_SIZE.get(),
@@ -266,7 +310,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                EndOresBlocks.REDSTONE_END_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        EndOresBlocks.REDSTONE_END_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.END_STONE.get(),
                 EndOresBlocks.REDSTONE_END_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.REDSTONE_THEEND_ORE_SIZE.get(),
@@ -281,7 +327,9 @@ public class OresFeatures {
      */
     public static void initNetherBasalt() {
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                BasaltOresBlocks.COAL_BASALT_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        BasaltOresBlocks.COAL_BASALT_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.BASALT.get(),
                 BasaltOresBlocks.COAL_BASALT_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.COAL_BASALT_ORE_SIZE.get(),
@@ -291,7 +339,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                BasaltOresBlocks.DIAMOND_BASALT_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        BasaltOresBlocks.DIAMOND_BASALT_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.BASALT.get(),
                 BasaltOresBlocks.DIAMOND_BASALT_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.DIAMOND_BASALT_ORE_SIZE.get(),
@@ -301,7 +351,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                BasaltOresBlocks.EMERALD_BASALT_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        BasaltOresBlocks.EMERALD_BASALT_ORE.getRegistryName()
+                ),
                 Blocks.BASALT.getDefaultState(),
                 BasaltOresBlocks.EMERALD_BASALT_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.EMERALD_BASALT_ORE_COUNT.get(),
@@ -310,7 +362,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                BasaltOresBlocks.GOLD_BASALT_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        BasaltOresBlocks.GOLD_BASALT_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.BASALT.get(),
                 BasaltOresBlocks.GOLD_BASALT_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.GOLD_BASALT_ORE_SIZE.get(),
@@ -320,7 +374,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                BasaltOresBlocks.IRON_BASALT_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        BasaltOresBlocks.IRON_BASALT_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.BASALT.get(),
                 BasaltOresBlocks.IRON_BASALT_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.IRON_BASALT_ORE_SIZE.get(),
@@ -330,7 +386,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                BasaltOresBlocks.LAPIS_BASALT_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        BasaltOresBlocks.LAPIS_BASALT_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.BASALT.get(),
                 BasaltOresBlocks.LAPIS_BASALT_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.LAPIS_BASALT_ORE_SIZE.get(),
@@ -340,7 +398,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                BasaltOresBlocks.REDSTONE_BASALT_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        BasaltOresBlocks.REDSTONE_BASALT_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.BASALT.get(),
                 BasaltOresBlocks.REDSTONE_BASALT_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.REDSTONE_BASALT_ORE_SIZE.get(),
@@ -355,7 +415,9 @@ public class OresFeatures {
      */
     public static void initNetherGravel() {
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                RegistryHelper.setPrefixOnRegistryName(GravelsOresBlocks.IRON_GRAVEL_ORE.getRegistryName(), "nether"),
+                Objects.requireNonNull(
+                        RegistryHelper.setPrefixOnRegistryName(GravelsOresBlocks.IRON_GRAVEL_ORE.getRegistryName(), "nether")
+                ),
                 ExtenedFillerBlockType.GRAVEL.get(),
                 GravelsOresBlocks.IRON_GRAVEL_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.NETHER_IRON_GRAVEL_ORE_SIZE.get(),
@@ -365,7 +427,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                RegistryHelper.setPrefixOnRegistryName(GravelsOresBlocks.GOLD_GRAVEL_ORE.getRegistryName(), "nether"),
+                Objects.requireNonNull(
+                        RegistryHelper.setPrefixOnRegistryName(GravelsOresBlocks.GOLD_GRAVEL_ORE.getRegistryName(), "nether")
+                ),
                 ExtenedFillerBlockType.GRAVEL.get(),
                 GravelsOresBlocks.GOLD_GRAVEL_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.NETHER_GOLD_GRAVEL_ORE_SIZE.get(),
@@ -375,7 +439,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                RegistryHelper.setPrefixOnRegistryName(GravelsOresBlocks.DIAMOND_GRAVEL_ORE.getRegistryName(), "nether"),
+                Objects.requireNonNull(
+                        RegistryHelper.setPrefixOnRegistryName(GravelsOresBlocks.DIAMOND_GRAVEL_ORE.getRegistryName(), "nether")
+                ),
                 ExtenedFillerBlockType.GRAVEL.get(),
                 GravelsOresBlocks.DIAMOND_GRAVEL_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.NETHER_DIAMOND_GRAVEL_ORE_SIZE.get(),
@@ -390,7 +456,9 @@ public class OresFeatures {
      */
     public static void initOceanGravel() {
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                RegistryHelper.setPrefixOnRegistryName(GravelsOresBlocks.IRON_GRAVEL_ORE.getRegistryName(), "ocean"),
+                Objects.requireNonNull(
+                        RegistryHelper.setPrefixOnRegistryName(GravelsOresBlocks.IRON_GRAVEL_ORE.getRegistryName(), "ocean")
+                ),
                 ExtenedFillerBlockType.GRAVEL.get(),
                 GravelsOresBlocks.IRON_GRAVEL_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.OCEAN_IRON_GRAVEL_ORE_SIZE.get(),
@@ -400,7 +468,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                RegistryHelper.setPrefixOnRegistryName(GravelsOresBlocks.GOLD_GRAVEL_ORE.getRegistryName(), "ocean"),
+                Objects.requireNonNull(
+                        RegistryHelper.setPrefixOnRegistryName(GravelsOresBlocks.GOLD_GRAVEL_ORE.getRegistryName(), "ocean")
+                ),
                 ExtenedFillerBlockType.GRAVEL.get(),
                 GravelsOresBlocks.GOLD_GRAVEL_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.OCEAN_GOLD_GRAVEL_ORE_SIZE.get(),
@@ -410,7 +480,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                RegistryHelper.setPrefixOnRegistryName(GravelsOresBlocks.DIAMOND_GRAVEL_ORE.getRegistryName(), "ocean"),
+                Objects.requireNonNull(
+                        RegistryHelper.setPrefixOnRegistryName(GravelsOresBlocks.DIAMOND_GRAVEL_ORE.getRegistryName(), "ocean")
+                ),
                 ExtenedFillerBlockType.GRAVEL.get(),
                 GravelsOresBlocks.DIAMOND_GRAVEL_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.OCEAN_DIAMOND_GRAVEL_ORE_SIZE.get(),
@@ -425,7 +497,9 @@ public class OresFeatures {
      */
     public static void initGravel() {
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                GravelsOresBlocks.IRON_GRAVEL_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        GravelsOresBlocks.IRON_GRAVEL_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.GRAVEL.get(),
                 GravelsOresBlocks.IRON_GRAVEL_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.IRON_GRAVEL_ORE_SIZE.get(),
@@ -435,7 +509,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                GravelsOresBlocks.GOLD_GRAVEL_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        GravelsOresBlocks.GOLD_GRAVEL_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.GRAVEL.get(),
                 GravelsOresBlocks.GOLD_GRAVEL_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.GOLD_GRAVEL_ORE_SIZE.get(),
@@ -445,7 +521,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                GravelsOresBlocks.DIAMOND_GRAVEL_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        GravelsOresBlocks.DIAMOND_GRAVEL_ORE.getRegistryName()
+                ),
                 ExtenedFillerBlockType.GRAVEL.get(),
                 GravelsOresBlocks.DIAMOND_GRAVEL_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.DIAMOND_GRAVEL_ORE_SIZE.get(),
@@ -460,7 +538,9 @@ public class OresFeatures {
      */
     public static void initNether() {
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.COAL_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.COAL_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.COAL_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.COAL_NETHER_ORE_SIZE.get(),
@@ -470,7 +550,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.DIAMOND_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.DIAMOND_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.DIAMOND_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.DIAMOND_NETHER_ORE_SIZE.get(),
@@ -480,7 +562,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerReplaceBlockFeature(
-                NetherOresBlocks.EMERALD_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.EMERALD_NETHER_ORE.getRegistryName()
+                ),
                 Blocks.NETHERRACK.getDefaultState(),
                 NetherOresBlocks.EMERALD_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.EMERALD_NETHER_ORE_COUNT.get(),
@@ -489,7 +573,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.GOLD_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.GOLD_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.GOLD_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.GOLD_NETHER_ORE_SIZE.get(),
@@ -499,7 +585,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.IRON_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.IRON_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.IRON_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.IRON_NETHER_ORE_SIZE.get(),
@@ -509,7 +597,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.LAPIS_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.LAPIS_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.LAPIS_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.LAPIS_NETHER_ORE_SIZE.get(),
@@ -519,7 +609,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.REDSTONE_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.REDSTONE_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.REDSTONE_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.REDSTONE_NETHER_ORE_SIZE.get(),
@@ -529,7 +621,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.ALUMINIUM_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.ALUMINIUM_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.ALUMINIUM_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.ALUMINIUM_NETHER_ORE_SIZE.get(),
@@ -539,7 +633,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.LEAD_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.LEAD_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.LEAD_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.LEAD_NETHER_ORE_SIZE.get(),
@@ -549,7 +645,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.TIN_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.TIN_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.TIN_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.TIN_NETHER_ORE_SIZE.get(),
@@ -559,7 +657,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.COPPER_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.COPPER_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.COPPER_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.COPPER_NETHER_ORE_SIZE.get(),
@@ -569,7 +669,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.RUBY_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.RUBY_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.RUBY_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.RUBY_NETHER_ORE_SIZE.get(),
@@ -579,7 +681,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.SAPPHIRE_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.SAPPHIRE_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.SAPPHIRE_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.SAPPHIRE_NETHER_ORE_SIZE.get(),
@@ -589,7 +693,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.SILVER_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.SILVER_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.SILVER_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.SILVER_NETHER_ORE_SIZE.get(),
@@ -599,7 +705,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.URANIUM_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.URANIUM_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.URANIUM_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.URANIUM_NETHER_ORE_SIZE.get(),
@@ -609,7 +717,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                NetherOresBlocks.OSMIUM_NETHER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        NetherOresBlocks.OSMIUM_NETHER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.NETHERRACK,
                 NetherOresBlocks.OSMIUM_NETHER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.OSMIUM_NETHER_ORE_SIZE.get(),
@@ -624,7 +734,9 @@ public class OresFeatures {
      */
     public static void initOverworld() {
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                OverworldOresBlocks.COPPER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        OverworldOresBlocks.COPPER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD,
                 OverworldOresBlocks.COPPER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.COPPER_ORE_SIZE.get(),
@@ -634,7 +746,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                OverworldOresBlocks.TIN_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        OverworldOresBlocks.TIN_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD,
                 OverworldOresBlocks.TIN_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.TIN_ORE_SIZE.get(),
@@ -644,7 +758,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                OverworldOresBlocks.ALUMINIUM_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        OverworldOresBlocks.ALUMINIUM_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD,
                 OverworldOresBlocks.ALUMINIUM_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.ALUMINIUM_ORE_SIZE.get(),
@@ -654,7 +770,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                OverworldOresBlocks.URANIUM_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        OverworldOresBlocks.URANIUM_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD,
                 OverworldOresBlocks.URANIUM_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.URANIUM_ORE_SIZE.get(),
@@ -664,7 +782,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                OverworldOresBlocks.OSMIUM_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        OverworldOresBlocks.OSMIUM_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD,
                 OverworldOresBlocks.OSMIUM_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.OSMIUM_ORE_SIZE.get(),
@@ -674,7 +794,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                OverworldOresBlocks.LEAD_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        OverworldOresBlocks.LEAD_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD,
                 OverworldOresBlocks.LEAD_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.LEAD_ORE_SIZE.get(),
@@ -684,7 +806,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                OverworldOresBlocks.SILVER_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        OverworldOresBlocks.SILVER_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD,
                 OverworldOresBlocks.SILVER_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.SILVER_ORE_SIZE.get(),
@@ -694,7 +818,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                OverworldOresBlocks.RUBY_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        OverworldOresBlocks.RUBY_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD,
                 OverworldOresBlocks.RUBY_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.RUBY_ORE_SIZE.get(),
@@ -704,7 +830,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                OverworldOresBlocks.SAPPHIRE_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        OverworldOresBlocks.SAPPHIRE_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD,
                 OverworldOresBlocks.SAPPHIRE_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.SAPPHIRE_ORE_SIZE.get(),
@@ -714,7 +842,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                OverworldOresBlocks.AMETHYST_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        OverworldOresBlocks.AMETHYST_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD,
                 OverworldOresBlocks.AMETHYST_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.AMETHYST_ORE_SIZE.get(),
@@ -724,7 +854,9 @@ public class OresFeatures {
         );
 
         WorldGenerationHelper.ConfiguredFeatureHelper.registerOreFeature(
-                OverworldOresBlocks.ZINC_ORE.getRegistryName(),
+                Objects.requireNonNull(
+                        OverworldOresBlocks.ZINC_ORE.getRegistryName()
+                ),
                 OreFeatureConfig.FillerBlockType.BASE_STONE_OVERWORLD,
                 OverworldOresBlocks.ZINC_ORE.getDefaultState(),
                 FREQUENCY_CONFIG.ZINC_ORE_SIZE.get(),